misc: separate create and open test
Previous test firstly create file and then try to open them.
Well, the open result could be wrong. After patch, we can verify
open separately and trustable.

To do this, we need to use a tool to put a test file on devices.
The tool is ${APK_REPO}/deck/touch-file-on-devices.sh, and it reads
the path list from ${APK_REPO}/jni/sample.path which is also used
by JNI methods in misc.cc