linker: add case "multiple library instance for different namespace"
Android Nougat introduced namespace based dynamic link which could be
taken as extension of Java ClassLoader at native world. As a result,
a user library has instance for every ClassLoader/namespace if it's
used. The *instance* here means a copy in memory of one same process.
This case is to verify such scenario.